+++ src/Symfony/Component/HttpFoundation/Session/Storage/Proxy/NativeProxy.php
@@ -11,12 +11,12 @@
 
 namespace Symfony\Component\HttpFoundation\Session\Storage\Proxy;
 
-@trigger_error('The '.__NAMESPACE__.'\NativeProxy class is deprecated since version 3.4 and will be removed in 4.0. Use the Symfony\Component\HttpFoundation\Session\Storage\Proxy\SessionHandlerProxy class instead.', E_USER_DEPRECATED);
+@trigger_error('The '.__NAMESPACE__.'\NativeProxy class is deprecated since version 3.4 and will be removed in 4.0. Use your session handler implementation directly.', E_USER_DEPRECATED);
 
 /**
  * This proxy is built-in session handlers in PHP 5.3.x.
  *
- * @deprecated since version 3.4, to be removed in 4.0. Use SessionHandlerProxy instead.
+ * @deprecated since version 3.4, to be removed in 4.0. Use your session handler implementation directly.
  *
  * @author Drak <drak@zikula.org>
  */